pyproject.toml
[project]
name = "tv"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "esptool>=5.0.2",
    "keyboard>=0.13.5",
    "pillow>=11.3.0",
    "pyserial>=3.5",
    "pystray>=0.19.5",
    "pywin32>=311",
    "wheel>=0.45.1",
    "pyinstaller>=6.15.0",
    "pynput>=1.8.1",
]
